requirements.txt
# Core dependencies
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
python-multipart>=0.0.5,<0.1.0
aiohttp>=3.8.0
aiofiles>=0.7.0,<0.8.0

# Security & Utilities
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=0.19.0

# Text extraction
pymupdf>=1.23.7           # PDF处理
pandas>=2.1.3             # 电子表格处理
openpyxl>=3.1.2          # Excel文件支持
xlrd>=2.0.1              # 旧版Excel支持
python-docx>=1.0.0       # Word文档处理
python-magic>=0.4.24,<0.5.0
python-magic-bin==0.4.14  # for Windows
chardet>=5.2.0           # 文本编码检测

# Storage
boto3>=1.26.0            # S3存储支持
botocore>=1.32.3
